refactor: move EventSubscriber from zenoh core to zenoh-ext with extension trait
Replace the inherent .event() impl on SubscriberBuilder in the zenoh
core crate with an EventSubscriberBuilderExt extension trait in
zenoh-ext, matching the AdvancedSubscriberBuilderExt pattern.

Closes #60
