+++ zenoh-ext/Cargo.toml
@@ -38,6 +38,7 @@ flume = { workspace = true }
 futures = { workspace = true }
 leb128 = { workspace = true }
 serde = { workspace = true, features = ["default"] }
+serde_json = { workspace = true }
 tokio = { workspace = true, features = [
   "io-std",
   "macros",
